Add basic CI via GitHub actions.
Additionally:

* Allow disabling setting up root password or sudo command wrangling
  (used in CI environment, but could be used by the end user via the
  MANAGE_SUDO environment variable).
* gnustep-web-install and gnustep-web-install-dev also detect if they
  are run inside a working copy, assume this is CI environment, and
  skip cloning the repo itself.
* Only gnustep-web-install (non-dev) is run in CI for now, as the
  dev version assumes ssh keys are set up to access private repos.
  This could be arranged in GitHub Actions, but is left for future
  improvement.

No attempt was made to cache any part of the installation between runs,
since we are trying to run the entire installation process as a test.
It would be useful to have a smoke test that just verifies portions of
the build, or to produce a custom container image that has at least
the base dependencies installed. Then the full run could be performed
on-demand. However, adjusting the script to run in such phases is
better left for a future improvement.

No attempt is done yet to improve failure wrangling (e.g. make the
scripts exit more often on errors, which is useful for CI).

No attempt is made to run on multiple OS versions or architectures.

Fixes #36.